能源

+++ YouFeel/Module/Applications/EnergyStatistics/EnergyStatisticBase/SubViews/EnergyTrend/Views/BXTEnergyTrendBudgetCell.m
@@ -28,7 +28,7 @@
     _energyTrendInfo = energyTrendInfo;
     
     self.budgeView.text = [NSString stringWithFormat:@"预算：%@ Kwh", energyTrendInfo.energy_consumption_budget];
-    self.differenceNumView.text = [NSString stringWithFormat:@"差异量：%ld Kwh", energyTrendInfo.energy_consumption_budget_diff];
+    self.differenceNumView.text = [NSString stringWithFormat:@"差异量：%ld Kwh", (long)energyTrendInfo.energy_consumption_budget_diff];
     self.differenceRateView.text = [NSString stringWithFormat:@"差异率：%@", energyTrendInfo.energy_consumption_budget_diff_per];
 }
 
@@ -37,7 +37,7 @@
     _moneyTrendInfo = moneyTrendInfo;
     
     self.budgeView.text = [NSString stringWithFormat:@"预算：%@ 元", moneyTrendInfo.money_budget];
-    self.differenceNumView.text = [NSString stringWithFormat:@"差异量：%ld 元", moneyTrendInfo.money_budget_diff];
+    self.differenceNumView.text = [NSString stringWithFormat:@"差异量：%ld 元", (long)moneyTrendInfo.money_budget_diff];
     self.differenceRateView.text = [NSString stringWithFormat:@"差异率：%@", moneyTrendInfo.money_budget_diff_per];
 }
