+++ YouFeel/Module/Applications/EnergyStatistics/EnergyStatisticBase/SubViews/EnergyTrend/Views/BXTEnergyTrendCell.m
@@ -27,8 +27,8 @@
 {
     _energyTrendInfo = energyTrendInfo;
     
-    self.consumptionView.text = [NSString stringWithFormat:@"总能耗：%ld Kwh", energyTrendInfo.energy_consumption];
-    self.unitConsumptionView.text = [NSString stringWithFormat:@"单位面积能耗：%ld Kwh", energyTrendInfo.energy_consumption_unit_area];
+    self.consumptionView.text = [NSString stringWithFormat:@"总能耗：%ld Kwh", (long)energyTrendInfo.energy_consumption];
+    self.unitConsumptionView.text = [NSString stringWithFormat:@"单位面积能耗：%ld Kwh", (long)energyTrendInfo.energy_consumption_unit_area];
     
     self.circleView.text = [NSString stringWithFormat:@"环比: %ld Kwh", (long)energyTrendInfo.mom_energy_consumption];
     self.circleImageView.image = [self judgeImage:energyTrendInfo.mom_energy_consumption_sign];
@@ -45,8 +45,8 @@
 {
     _moneyTrendInfo = moneyTrendInfo;
     
-    self.consumptionView.text = [NSString stringWithFormat:@"总费用：%ld 元", moneyTrendInfo.money];
-    self.unitConsumptionView.text = [NSString stringWithFormat:@"单位面积费用：%ld 元", moneyTrendInfo.money_unit_area];
+    self.consumptionView.text = [NSString stringWithFormat:@"总费用：%ld 元", (long)moneyTrendInfo.money];
+    self.unitConsumptionView.text = [NSString stringWithFormat:@"单位面积费用：%ld 元", (long)moneyTrendInfo.money_unit_area];
     
     self.circleView.text = [NSString stringWithFormat:@"环比: %ld 元", (long)moneyTrendInfo.mom_money];
     self.circleImageView.image = [self judgeImage:moneyTrendInfo.mom_money_sign];
